Modify behaviour of Product Detail Serializer and other minor fixes
In order to create new categories when a new product is being created, overriding the seralizer
`.create()` method to create new or get an old category and assign to the product.

Added a new permission `IsStoreStaffAdminUser` to allow only store staff users to create new products/categories
and everyone to get all the products/categories.